fix: GCS blobs signed URLS generation
Add a workaround for the inability to sign GCS blob URLs when running on Google Cloud Run. See [here](https://stackoverflow.com/questions/64234214/how-to-generate-a-blob-signed-url-in-google-cloud-run) for details. The workaround involves loading a service account JSON key file content's from an environment variable and using that to explicitly authenticate Google APIs/SDKs.

Included also are other minor improvements and fixes.